+++ src/prelude.rs
@@ -2,4 +2,6 @@ pub use crate::goose::{
     GooseTask, GooseTaskError, GooseTaskFunction, GooseTaskResult, GooseTaskSet, GooseUser,
 };
 pub use crate::metrics::GooseMetrics;
-pub use crate::{task, taskset, GooseAttack, GooseDefault, GooseDefaultType, GooseError};
+pub use crate::{
+    task, taskset, GooseAttack, GooseDefault, GooseDefaultType, GooseError, GooseTaskSetScheduler,
+};